+++ projects/start-sdk/lib/index.ts
@@ -43,6 +43,7 @@ export {
 
 export * as actions from '@start9labs/start-core/actions'
 export * as backup from './backup'
+export * as primaryUrl from './primaryUrl'
 export * as daemons from './mainFn/Daemons'
 export * as health from './health'
 export * as healthFns from './health/checkFns'